test: consolidate network tests in one file
